+++ v3/pkg/application/webview_window.go
@@ -44,7 +44,6 @@ type (
 		size() (int, int)
 		width() int
 		height() int
-		relativePosition() (int, int)
 		destroy()
 		reload()
 		forceReload()
@@ -57,7 +56,6 @@ type (
 		close()
 		zoom()
 		setHTML(html string)
-		setRelativePosition(x int, y int)
 		on(eventID uint)
 		minimise()
 		unminimise()
@@ -82,8 +80,14 @@ type (
 		startResize(border string) error
 		print() error
 		setEnabled(enabled bool)
+		physicalBounds() Rect
+		setPhysicalBounds(physicalBounds Rect)
+		bounds() Rect
+		setBounds(bounds Rect)
 		position() (int, int)
 		setPosition(x int, y int)
+		relativePosition() (int, int)
+		setRelativePosition(x int, y int)
 		flash(enabled bool)
 		handleKeyEvent(acceleratorString string)
 		getBorderSizes() *LRTB
@@ -782,7 +786,73 @@ func (w *WebviewWindow) Height() int {
 	return InvokeSyncWithResult(w.impl.height)
 }
 
-// RelativePosition returns the relative position of the window to the screen
+// PhysicalBounds returns the physical bounds of the window
+func (w *WebviewWindow) PhysicalBounds() Rect {
+	if w.impl == nil && !w.isDestroyed() {
+		return Rect{}
+	}
+	var rect Rect
+	InvokeSync(func() {
+		rect = w.impl.physicalBounds()
+	})
+	return rect
+}
+
+// SetPhysicalBounds sets the physical bounds of the window
+func (w *WebviewWindow) SetPhysicalBounds(physicalBounds Rect) {
+	if w.impl == nil && !w.isDestroyed() {
+		return
+	}
+	InvokeSync(func() {
+		w.impl.setPhysicalBounds(physicalBounds)
+	})
+}
+
+// Bounds returns the DIP bounds of the window
+func (w *WebviewWindow) Bounds() Rect {
+	if w.impl == nil && !w.isDestroyed() {
+		return Rect{}
+	}
+	var rect Rect
+	InvokeSync(func() {
+		rect = w.impl.bounds()
+	})
+	return rect
+}
+
+// SetBounds sets the DIP bounds of the window
+func (w *WebviewWindow) SetBounds(bounds Rect) {
+	if w.impl == nil && !w.isDestroyed() {
+		return
+	}
+	InvokeSync(func() {
+		w.impl.setBounds(bounds)
+	})
+}
+
+// Position returns the absolute position of the window
+func (w *WebviewWindow) Position() (int, int) {
+	if w.impl == nil && !w.isDestroyed() {
+		return 0, 0
+	}
+	var x, y int
+	InvokeSync(func() {
+		x, y = w.impl.position()
+	})
+	return x, y
+}
+
+// SetPosition sets the absolute position of the window.
+func (w *WebviewWindow) SetPosition(x int, y int) {
+	if w.impl == nil && !w.isDestroyed() {
+		return
+	}
+	InvokeSync(func() {
+		w.impl.setPosition(x, y)
+	})
+}
+
+// RelativePosition returns the position of the window relative to the screen WorkArea on which it is
 func (w *WebviewWindow) RelativePosition() (int, int) {
 	if w.impl == nil && !w.isDestroyed() {
 		return 0, 0
@@ -794,16 +864,16 @@ func (w *WebviewWindow) RelativePosition() (int, int) {
 	return x, y
 }
 
-// Position returns the absolute position of the window to the screen
-func (w *WebviewWindow) Position() (int, int) {
-	if w.impl == nil && !w.isDestroyed() {
-		return 0, 0
+// SetRelativePosition sets the position of the window relative to the screen WorkArea on which it is.
+func (w *WebviewWindow) SetRelativePosition(x, y int) Window {
+	w.options.X = x
+	w.options.Y = y
+	if w.impl != nil {
+		InvokeSync(func() {
+			w.impl.setRelativePosition(x, y)
+		})
 	}
-	var x, y int
-	InvokeSync(func() {
-		x, y = w.impl.position()
-	})
-	return x, y
+	return w
 }
 
 func (w *WebviewWindow) Destroy() {
@@ -921,18 +991,6 @@ func (w *WebviewWindow) SetHTML(html string) Window {
 	if w.impl != nil {
 		InvokeSync(func() {
 			w.impl.setHTML(html)
-		})
-	}
-	return w
-}
-
-// SetRelativePosition sets the position of the window.
-func (w *WebviewWindow) SetRelativePosition(x, y int) Window {
-	w.options.X = x
-	w.options.Y = y
-	if w.impl != nil {
-		InvokeSync(func() {
-			w.impl.setRelativePosition(x, y)
 		})
 	}
 	return w
@@ -1168,16 +1226,6 @@ func (w *WebviewWindow) SetEnabled(enabled bool) {
 	}
 	InvokeSync(func() {
 		w.impl.setEnabled(enabled)
-	})
-}
-
-func (w *WebviewWindow) SetPosition(x int, y int) {
-	// set absolute position
-	if w.impl == nil && !w.isDestroyed() {
-		return
-	}
-	InvokeSync(func() {
-		w.impl.setPosition(x, y)
 	})
 }